Changed indexing from 'int' to 'size_t'. If '.vf' compilation fails revert to direct '.vf' loading.

+++ knut/demo/logistic/sys-logistic.cpp
@@ -9,9 +9,9 @@
 extern "C"
 {
 
-int sys_ndim(){ return 1; }
-int sys_npar(){ return 2; }
-int sys_ntau(){ return 2; }
+size_t sys_ndim(){ return 1; }
+size_t sys_npar(){ return 2; }
+size_t sys_ntau(){ return 2; }
 
 void sys_tau( KNVector& out, double t, const KNVector& par )
 {
@@ -19,7 +19,7 @@ void sys_tau( KNVector& out, double t, const KNVector& par )
   out(1) = 1.0;
 }
 
-void sys_dtau( KNVector& out, double t, const KNVector& par, int vp )
+void sys_dtau( KNVector& out, double t, const KNVector& par, size_t vp )
 {
 	switch( vp )
 	{
@@ -43,7 +43,7 @@ void sys_rhs( KNVector& out, double t, const KNMatrix& x, const KNVector& par )
 }
 
 void sys_deri( KNMatrix &out, double t, const KNMatrix& x, const KNVector& par, 
-	       int nx, const int* vx, int np, const int* vp, const KNMatrix& vv )
+	       size_t nx, const size_t* vx, size_t np, const size_t* vp, const KNMatrix& vv )
 {
   if( (t < 0)||(t > 1) ){ std::cout << "deri: t is not element of the interval\n"; }
