+++ knut/demo/traffic/sys-traffic3cars.cpp
@@ -27,23 +27,23 @@
 extern "C"
 {
 
-int sys_ndim() { return 5; }
-int sys_npar() { return 5; }
-int sys_ntau() { return 2; }
-int sys_nderi() { return 0; }
+size_t sys_ndim() { return 5; }
+size_t sys_npar() { return 5; }
+size_t sys_ntau() { return 2; }
+size_t sys_nderi() { return 0; }
 
 void sys_p_tau( KNArray2D<double>& out, const KNArray1D<double>& time, const KNArray1D<double>& par )
 {
-  for (int idx = 0; idx < time.size(); ++idx)
+  for (size_t idx = 0; idx < time.size(); ++idx)
   {
     out(0,idx) = 0.0;
     out(1,idx) = par(2);
   }
 }
 
-void sys_p_dtau( KNArray2D<double>& out, const KNArray1D<double>& time, const KNArray1D<double>& par, int vp )
+void sys_p_dtau( KNArray2D<double>& out, const KNArray1D<double>& time, const KNArray1D<double>& par, size_t vp )
 {
-  for (int idx = 0; idx < time.size(); ++idx)
+  for (size_t idx = 0; idx < time.size(); ++idx)
   {
     out(0,idx) = 0.0;
     if( vp == 2 ) out(1,idx) = 1.0;
@@ -56,11 +56,11 @@ static inline double V( double x )
   return pow( ( x - 1 ) / 1, 3.0 ) / ( pow( ( x - 1 ) / 1, 3.0 ) + 1);
 }
 
-void sys_p_rhs( KNArray2D<double>& out, const KNArray1D<double>& time, const KNArray3D<double>& yy, const KNArray1D<double>& par, int sel )
+void sys_p_rhs( KNArray2D<double>& out, const KNArray1D<double>& time, const KNArray3D<double>& yy, const KNArray1D<double>& par, size_t sel )
 {
 #define xx(i,j) yy(i-1,j-1,idx)
 #define f(i,j) out(i-1,idx)
-  for (int idx = 0; idx < time.size(); ++idx)
+  for (size_t idx = 0; idx < time.size(); ++idx)
   {
     if ( xx(4,2) > 1 ){
       f(1,1) = par(3) * ( par(4) * V(xx(4,2)) - xx(1,1) ); 
@@ -69,7 +69,7 @@ void sys_p_rhs( KNArray2D<double>& out, const KNArray1D<double>& time, const KNA
     }
   }
 
-  for (int idx = 0; idx < time.size(); ++idx)
+  for (size_t idx = 0; idx < time.size(); ++idx)
   {
     if( xx(5,2) > 1 ){
       f(2,1) = par(3) * ( par(4) * V(xx(5,2)) - xx(2,1) ); 
@@ -78,7 +78,7 @@ void sys_p_rhs( KNArray2D<double>& out, const KNArray1D<double>& time, const KNA
     }
   }
 
-  for (int idx = 0; idx < time.size(); ++idx)
+  for (size_t idx = 0; idx < time.size(); ++idx)
   {
     if( ( 3*par(1) - xx(5,2) - xx(4,2) ) > 1 ){
       f(3,1) = par(3) * ( par(4) * V( 3*par(1) - xx(5,2) - xx(4,2) ) - xx(3,1) );
@@ -87,15 +87,15 @@ void sys_p_rhs( KNArray2D<double>& out, const KNArray1D<double>& time, const KNA
     }
   }
 
-  for (int idx = 0; idx < time.size(); ++idx)
+  for (size_t idx = 0; idx < time.size(); ++idx)
   {
     f(4,1) = xx(2,1) - xx(1,1);
     f(5,1) = xx(3,1) - xx(2,1);
   }
 }
 
-void sys_p_deri( KNArray3D<double>&, const KNArray1D<double>& time, const KNArray3D<double>& yy, const KNArray1D<double>& par, int sel, 
-                 int nx, const int* vx, int np, const int* vp, const KNArray3D<double>& vv )
+void sys_p_deri( KNArray3D<double>&, const KNArray1D<double>& time, const KNArray3D<double>& yy, const KNArray1D<double>& par, size_t sel, 
+                 size_t nx, const size_t* vx, size_t np, const size_t* vp, const KNArray3D<double>& vv )
 {
 
 }
