+++ knut/demo/turn/sys-turn.cpp
@@ -33,10 +33,10 @@ static inline double cf( double x )
 extern "C"
 {
 
-int sys_ndim(){ return 2; }
-int sys_npar(){ return 3; }
-int sys_ntau(){ return 2; }
-int sys_nderi(){ return 0; }
+size_t sys_ndim(){ return 2; }
+size_t sys_npar(){ return 3; }
+size_t sys_ntau(){ return 2; }
+size_t sys_nderi(){ return 0; }
 
 
 void sys_tau( KNVector& out, double t, const KNVector& par )
@@ -45,7 +45,7 @@ void sys_tau( KNVector& out, double t, const KNVector& par )
   out(1) = par(2);
 }
 
-void sys_dtau( KNVector& out, double t, const KNVector& par, int vp )
+void sys_dtau( KNVector& out, double t, const KNVector& par, size_t vp )
 {
 	out(0) = 0.0;
 	if( vp == 2 ) out(1) = 1.0;
@@ -59,7 +59,7 @@ void sys_rhs( KNVector& out, double t, const KNMatrix& x, const KNVector& par )
 }
 
 void sys_deri( KNMatrix &out, double t, const KNMatrix& x, const KNVector& par, 
-	       int nx, const int* vx, int np, const int* vp, const KNMatrix& vv )
+	       size_t nx, const size_t* vx, size_t np, const size_t* vp, const KNMatrix& vv )
 {
 
 }
